requirements.txt
fastapi
uvicorn[standard]
google-genai
pydantic
python-dotenv
aiosqlite
pytest
pytest_asyncio